refactor(Client): refactor client code into a single class
Instead of having two Clients (one with an async constructor, one with a sync
one), have a single client that exposes a async `loadSpec` method that
essentially implements the async constructor.
