+++ examples/sync-client-version.js
@@ -2,7 +2,7 @@
 // Create an API client based on a specified API version. kubernetes-client uses
 // included swagger specification files.
 //
-const Client = require('kubernetes-client').SyncClient
+const Client = require('kubernetes-client').Client
 const config = require('kubernetes-client').config;
 
 const deploymentManifest = require('./nginx-deployment.json');
