+++ lib/index.js
@@ -19,6 +19,5 @@ module.exports.testUtils = {
 //
 // Current interface
 //
-module.exports.SyncClient = require('./sync-client');
-module.exports.Client = require('./client');
+module.exports.Client = require('./swagger-client');
 module.exports.config = require('./config');
