+++ lib/swagger-client.js
@@ -1,4 +1,5 @@
 'use strict';
+/* eslint-disable no-sync */
 
 /**
  * @file Convert a Swagger specification into a kubernetes-client API client.
@@ -22,9 +23,14 @@
  * [3]: https://swagger.io/specification/#pathsObject
  */
 
+const fs = require('fs');
+const path = require('path');
+const zlib = require('zlib');
+
 const merge = require('lodash.merge');
 
 const getAliases = require('./common').getAliases;
+const Request = require('./request');
 
 class Endpoint {
   /**
@@ -91,6 +97,23 @@ class Component {
     component.splits = options.splits.slice();
     component.http = options.http;
     return component;
+  }
+
+  /**
+   * Load swagger.json from kube-apiserver specified via this.http.
+   * @returns {Promise} Promise
+   */
+  loadSpec() {
+    return new Promise((resolve, reject) => {
+      this.http.request('GET', { path: '/swagger.json' }, (err, res) => {
+        if (err) return reject(err);
+        if (res.statusCode !== 200) {
+          return reject(new Error(`Failed to get swagger.json: ${ res.statusCode }`));
+        }
+        this._addSpec(res.body);
+        resolve(this);
+      });
+    });
   }
 
   /**
@@ -308,21 +331,22 @@ class Component {
   }
 }
 
-class SwaggerClient {
-
-  /**
-   * Swagger-based API client.
-   *
-   * @param {object} options - Request options object.
-   * @param {object} options.spec - Swagger specification
-   * @param {object} options.http - kubernetes-client Request object
-   */
+class Client {
   constructor(options) {
-    const spec = options.spec;
-    const root = new Component({ splits: [], http: options.http });
-    root._addSpec(spec);
+    const http = options.http || new Request(options.config);
+    let spec = options.spec;
+    if (!spec && options.version) {
+      const swaggerPath = path.join(
+        __dirname,
+        'specs',
+        `swagger-${ options.version }.json.gz`);
+      spec = JSON.parse(zlib.gunzipSync(fs.readFileSync(swaggerPath)));
+    }
+
+    const root = new Component({ splits: [], http: http });
+    if (spec) root._addSpec(spec);
     return root;
   }
 }
 
-module.exports = SwaggerClient;
+module.exports = Client;
