+++ test/swagger-client.test.js
@@ -1,94 +1,153 @@
+/* eslint-disable max-nested-callbacks */
 'use strict';
 
 const assume = require('assume');
-const SwaggerClient = require('../lib/swagger-client');
+const nock = require('nock');
 
-describe('lib.swagger', () => {
-  describe('SwaggerClient', () => {
-    it('expands Paths Object', () => {
-      const spec = {
-        paths: {
-          '/foo/bar/': { },
-          'baz/zab': { }
-        }
-      };
-      const client = new SwaggerClient({ spec });
-      assume(client.foo).is.truthy();
-      assume(client.foo.bar).is.truthy();
-      assume(client.baz).is.truthy();
-      assume(client.baz.zab).is.truthy();
+const common = require('./common');
+const Client = require('../lib/swagger-client');
+
+const beforeTesting = common.beforeTesting;
+
+describe('lib.swagger-client', () => {
+  describe('.Client', () => {
+
+    describe('.loadSpec', () => {
+      beforeTesting('unit', () => {
+        nock(common.api.url)
+          .get('/swagger.json')
+          .reply(200, {
+            paths: {
+              '/api/': {
+                get: {
+                  operationId: 'getCoreAPIVersions'
+                }
+              }
+            }
+          });
+      });
+
+      it('creates a dynamically generated client', done => {
+        const config = { url: common.api.url };
+        const client = new Client({ config });
+        client.loadSpec()
+          .then(() => {
+            assume(client.api.get).is.a('function');
+            done();
+          })
+          .catch(err => done(err));
+      });
     });
 
-    it('adds operations defined by a Path Item Object', () => {
-      const spec = {
-        paths: {
-          '/foo/bar/': {
-            get: {
-              operationId: 'fooBarGet'
+    describe('.constructor', () => {
+      it('creates a dynamically generated client synchronously based on version', () => {
+        const options = { config: {}, version: '1.9' };
+        const client = new Client(options);
+        assume(client.api.get).is.a('function');
+      });
+
+      it('creates a dynamically generated client synchronously from swagger spec', () => {
+        const options = {
+          config: {},
+          spec: {
+            paths: {
+              '/api/': {
+                get: {
+                  operationId: 'getCoreAPIVersions'
+                }
+              }
             }
           }
-        }
-      };
-      const client = new SwaggerClient({ spec });
-      assume(client.foo.bar.get).is.a('function');
-    });
+        };
+        const client = new Client(options);
+        assume(client.api.get).is.a('function');
+      });
 
-    it('represents Path Templating with functions', () => {
-      const spec = {
-        paths: {
-          '/foo/{name}/bar': { },
-          '/foo': {
-            get: {
-              operationId: 'fooGet'
+      it('expands Paths Object', () => {
+        const spec = {
+          paths: {
+            '/foo/bar/': { },
+            'baz/zab': { }
+          }
+        };
+        const client = new Client({ spec, http: {}});
+        assume(client.foo).is.truthy();
+        assume(client.foo.bar).is.truthy();
+        assume(client.baz).is.truthy();
+        assume(client.baz.zab).is.truthy();
+      });
+
+      it('adds operations defined by a Path Item Object', () => {
+        const spec = {
+          paths: {
+            '/foo/bar/': {
+              get: {
+                operationId: 'fooBarGet'
+              }
             }
           }
-        }
-      };
-      const client = new SwaggerClient({ spec });
-      assume(client.foo).is.truthy();
-      assume(client.foo.get).is.a('function');
-      assume(client.foo.getStream).is.a('function');
+        };
+        const client = new Client({ spec, http: {}});
+        assume(client.foo.bar.get).is.a('function');
+      });
 
-      assume(client.foo.bar).is.falsy();
-      assume(client.foo).is.a('function');
-      assume(client.foo('zoo').bar).is.truthy();
-    });
-
-    it('aliases resources', () => {
-      const spec = {
-        paths: {
-          '/foo/deployments': {
-            get: {
-              operationId: 'fooDeploymentsGet'
+      it('represents Path Templating with functions', () => {
+        const spec = {
+          paths: {
+            '/foo/{name}/bar': { },
+            '/foo': {
+              get: {
+                operationId: 'fooGet'
+              }
             }
           }
-        }
-      };
-      const client = new SwaggerClient({ spec });
-      assume(client.foo.deployments).is.truthy();
-      assume(client.foo.deployment).is.truthy();
-      assume(client.foo.deploy).is.truthy();
-    });
+        };
+        const client = new Client({ spec, http: {}});
+        assume(client.foo).is.truthy();
+        assume(client.foo.get).is.a('function');
+        assume(client.foo.getStream).is.a('function');
 
-    it('adds functions for Namespaced CustomResourceDefinitions', () => {
-      const client = new SwaggerClient({ spec: { paths: {}}});
-      const crd = {
-        spec: {
-          group: 'stable.example.com',
-          version: 'v1',
-          names: {
-            plural: 'foos'
+        assume(client.foo.bar).is.falsy();
+        assume(client.foo).is.a('function');
+        assume(client.foo('zoo').bar).is.truthy();
+      });
+
+      it('aliases resources', () => {
+        const spec = {
+          paths: {
+            '/foo/deployments': {
+              get: {
+                operationId: 'fooDeploymentsGet'
+              }
+            }
           }
-        }
-      };
-      client.addCustomResourceDefinition(crd);
-      assume(client.apis['stable.example.com'].v1.namespaces('default').foos.get).is.a('function');
-      assume(client.apis['stable.example.com'].v1.namespaces('default').foos('blah').get).is.a('function');
-      assume(client.apis['stable.example.com'].v1.namespaces('default').foos('blah').delete).is.a('function');
-      assume(client.apis['stable.example.com'].v1.namespaces('default').foos('blah').get).is.a('function');
-      assume(client.apis['stable.example.com'].v1.namespaces('default').foos('blah').patch).is.a('function');
-      assume(client.apis['stable.example.com'].v1.namespaces('default').foos('blah').post).is.a('function');
-      assume(client.apis['stable.example.com'].v1.namespaces('default').foos('blah').put).is.a('function');
+        };
+        const client = new Client({ spec, http: {}});
+        assume(client.foo.deployments).is.truthy();
+        assume(client.foo.deployment).is.truthy();
+        assume(client.foo.deploy).is.truthy();
+      });
+
+      it('adds functions for Namespaced CustomResourceDefinitions', () => {
+        const client = new Client({ spec: { paths: {}}, http: {}});
+        const crd = {
+          spec: {
+            group: 'stable.example.com',
+            version: 'v1',
+            names: {
+              plural: 'foos'
+            }
+          }
+        };
+        client.addCustomResourceDefinition(crd);
+        assume(client.apis['stable.example.com'].v1.namespaces('default').foos.get).is.a('function');
+        assume(client.apis['stable.example.com'].v1.namespaces('default').foos('blah').get).is.a('function');
+        assume(client.apis['stable.example.com'].v1.namespaces('default').foos('blah').delete).is.a('function');
+        assume(client.apis['stable.example.com'].v1.namespaces('default').foos('blah').get).is.a('function');
+        assume(client.apis['stable.example.com'].v1.namespaces('default').foos('blah').patch).is.a('function');
+        assume(client.apis['stable.example.com'].v1.namespaces('default').foos('blah').post).is.a('function');
+        assume(client.apis['stable.example.com'].v1.namespaces('default').foos('blah').put).is.a('function');
+      });
     });
   });
 });
